tidying up js, removing repeated code, using requirejs
