+++ inc/html/admin/sources.php
@@ -5,23 +5,13 @@ if (!defined('CONFIG_PROTECTION')) {
     exit;
 }
 ?>
-<script>
-    $(document).ready(function() {
-        $("#search-area-admin").on("keyup", function() {
-            var value = $(this).val().toLowerCase();
-            $(".table tbody tr").filter(function() {
-                $(this).toggle($(this).text().toLowerCase().indexOf(value) > -1)
-            });
-        });
-    });
-</script>
 <div class="col-xs-12 col-sm-12 col-lg-12 ">
     <nav>
         <ol class="breadcrumb bg-dark text-light">
             <li class="breadcrumb-item ">
                 <a href="admin.php">Admin</a>
             </li>
-            <li class="breadcrumb-item active text-light">
+            <li id="sources-page" class="breadcrumb-item active text-light">
                 Sources
             </li>
         </ol>
@@ -30,7 +20,7 @@ if (!defined('CONFIG_PROTECTION')) {
         <div class="input-group-prepend">
             <span class="input-group-text text-light bg-dark" id="basic-addon1">Search</span>
         </div>
-        <input id="search-area-admin" type="text" class="form-control text-light bg-dark" placeholder="Example: chiuni" aria-label="Search" aria-describedby="basic-addon1">
+        <input id="search-area-admin" type="text" class="form-control text-light bg-dark search-me" placeholder="Example: chiuni" aria-label="Search" aria-describedby="basic-addon1">
     </div>
     <div class="table-responsive" style="max-height: 65vh;">
         <table class="table table-dark tableFixHead table-striped " data-sortable>
@@ -93,8 +83,3 @@ if (!defined('CONFIG_PROTECTION')) {
         </button>
     </div>
 </div>
-<script>
-    $('.add-source .btn .btn-success').click(function() {
-        $('form[name="modalForm"]').submit();
-    });
-</script>
