+++ js/.htaccess
@@ -1,0 +1,1 @@
+Deny From All
