+++ js/main.js
@@ -3,9 +3,9 @@ define([
     'methods',
     'modal',
     'feed',
+    'admin',
     'bootstrap',
-    'toggle',
-    'lodash'
+    'toggle'
 ], function ($, methods, modal, feed) {
     $("body").tooltip({
         selector: '[data-toggle=tooltip]'
@@ -16,13 +16,18 @@ define([
     }
     var username = methods.getUserName();
     var safelock = methods.getSafeLock();
-    console.log(username)
+    // console.log(username)
 
     $(".search-me").on("keyup", methods.searchOnKeyUp);
     if ($("#news-feed").length) {
         feed.loadMore(username, safelock);
     };
+    //Login page js, toggles new account creation mode
+    if ($("#newaccount").length) {
+        $('#newaccount').click(function () {
+            $('#name-input').toggle();
+        });
+    };
     modal.refreshSubscribed(username, safelock, 'topics');
     modal.refreshSubscribed(username, safelock, 'pages');
-
 });
