--- js/search.js
@@ -1,8 +1,0 @@
-function searchOnKeyUp(id, field) {
-    $("#" + id).on("keyup", function () {
-        var value = $(this).val().toLowerCase();
-        $("." + field).filter(function () {
-            $(this).toggle($(this).text().toLowerCase().indexOf(value) > -1)
-        });
-    });
-}
